test(networking): move ConnectionState/Direction tests to mirrored test_types.py
The ConnectionState and Direction enums are defined in
node/networking/types.py, so their tests belong in the mirrored
test_types.py, not test_peer.py.

Move TestConnectionState and TestDirection into a new
tests/node/networking/test_types.py. Leave the PeerInfo tests in
test_peer.py, which still use both enums and Multiaddr.
